V "1.0.0.2
Updated Common files
Circuit Diagram
TestJsonWiFi.txt

+++ MQTT bridge/common/driver/include/wifi.h
@@ -13,7 +13,7 @@
 typedef void (*InitWiFiCb_t)(void);
 typedef void (*WifiCallback)(uint8_t);
 void WIFI_Connect(uint8_t* ssid, uint8_t* pass, uint8_t* deviceName, WifiCallback cb);
-uint16 WIFI_ConnectTime(void);
+uint32 WIFI_ConnectTime(void);
 char *getBestSSID(void);
 STATUS WiFiScanStatus(void);
 uint8 WiFiConnectStatus(void);
